pyproject.toml
[tool.poetry]
name = "ESMDisPred"
version = "0.1.0"
description = ""
authors = ["Md Wasi Ul Kabir <mkabir3@uno.edu>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
numpy = "^1.26.3"
pandas = "^2.1.4"
ipykernel = "^6.28.0"
pyarrow = "^15.0.0"
biopython = "^1.83"
scikit-learn = "^1.4.1"
ipywidgets = "^8.1.2"
lightgbm = "^4.3.0"
torch = "^2.2.1"
fair-esm = "^2.0.0"
mlflow = "^2.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
